+++ CHANGELOG.md
@@ -16,6 +16,22 @@
 - Added **Lists Tweet Lookup** endpoint. ([#26](https://github.com/twitter-dart/twitter-api-v2/issues/26))
   - GET /2/lists/:id/tweets
 
+---
+
+- Refactored method names. Prefixes for methods that perform `GET` communication other than the search endpoint have been unified with `lookup`. This makes it a destructive change, but please refer to the following correspondence table for details of the change.
+  - **Tweets Service**
+    - From `likingUsers` to `lookupLikingUsers`
+    - From `likingTweets` to `lookupLikedTweets`
+    - From `retweetedBy` to `lookupRetweetedUsers`
+    - From `quoteTweets` to `lookupQuoteTweets`
+    - From `bookmarks` to `lookupBookmarks`
+  - **Users Service**
+    - From `followers` to `lookupFollowers`
+    - From `followings` to `lookupFollowings`
+    - From `mutingUsers` to `lookupMutingUsers`
+  - **Lists Service**
+    - From `pinnedLists` to `lookupPinnedLists`
+
 ## v1.6.0
 
 - Fixed a bug in `DELETE` communication.